backend/requirements.txt

Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
stripe==12.4.0
django-cors-headers==4.0.0
